moved more files around and moved remained of data_querier to transformer.rs files and was able to break down some more complex code into smaller chunks
