+++ src/datastore/sqlite.rs
@@ -318,34 +318,3 @@ pub(crate) async fn clear_sqlite_data(){
     }
 }
 
-pub fn sqlite_row_to_string_vec(row:&SqliteRow, log: &Log) -> Vec<String> {
-    // convert sqliteRow to csv row
-    let mut csv_row = Vec::new();
-    let number_of_columns = row.columns().len();
-    for i in 0..number_of_columns {
-        let column_type = row.column(i).type_info().to_string();
-        match column_type.as_str() {
-            "TEXT" => {
-                let value: String = row.get(i);
-                csv_row.push(value);
-            }
-            "INTEGER" => {
-                let value: i64 = row.get(i);
-                csv_row.push(value.to_string());
-            }
-            "REAL" => {
-                let value: f64 = row.get(i);
-                csv_row.push(value.to_string());
-            }
-            "BLOB" => {
-                let value: String = row.get(i);
-                csv_row.push(value);
-            }
-            _ => {
-                log.error(&format!("unknown column type: {}", column_type));
-            }
-        }
-    }
-    // finally return csv row
-    csv_row
-}
